[Fix] Translate pubsub message to dict early
Fixed an issue where a PSMessage instance was treated as a dict.
The easiest fix is to translate this object to a dictionary to
maintain compatibility.
