+++ src/aleph/services/utils.py
@@ -1,8 +1,10 @@
 import logging
 import re
 import socket
+from typing import Any, Dict
 
 import aiohttp
+from p2pclient.pb.p2pd_pb2 import PSMessage
 
 logger = logging.getLogger(__name__)
 
@@ -46,3 +48,18 @@ async def get_IP() -> str:
     except Exception as error:
         logging.exception("Error when fetching IPv4 from service")
         return get_ip4_from_socket()
+
+
+def pubsub_msg_to_dict(pubsub_msg: PSMessage) -> Dict[str, Any]:
+    """
+    A compatibility method that translates a p2pd pubsub message to the equivalent dictionary.
+    The returned value is then passed to `handle_incoming_host`.
+
+    TODO: use a better system than a dict to pass these parameters around.
+    """
+    return {
+        "from": getattr(pubsub_msg, "from"),
+        "data": pubsub_msg.data,
+        "seqno": pubsub_msg.seqno,
+        "topicIDs": pubsub_msg.topicIDs,
+    }
